+++ driver-core/src/test/java/com/datastax/driver/core/querybuilder/QueryBuilderExecutionTest.java
@@ -22,10 +22,7 @@ import java.util.List;
 
 import org.testng.annotations.Test;
 
-import com.datastax.driver.core.CCMBridge;
-import com.datastax.driver.core.RegularStatement;
-import com.datastax.driver.core.Row;
-import com.datastax.driver.core.TestUtils;
+import com.datastax.driver.core.*;
 
 import static com.datastax.driver.core.querybuilder.QueryBuilder.*;
 import static org.testng.Assert.*;
@@ -84,5 +81,24 @@ public class QueryBuilderExecutionTest extends CCMBridge.PerClassSingleNodeClust
         String query = "INSERT INTO foo(a,b,c,d) VALUES ('foo','bar',?,0);";
         RegularStatement stmt = insertInto("foo").value("a", "foo").value("b", "bar").value("c", bindMarker()).value("d", 0);
         assertEquals(query, stmt.getQueryString());
+    }
+
+    @Test(groups = "short")
+    public void batchNonBuiltStatementTest() throws Exception {
+
+        SimpleStatement simple = new SimpleStatement("INSERT INTO " + TABLE1 + " (k, t) VALUES (?, ?)", "batchTest1", "val1");
+        RegularStatement built = insertInto(TABLE1).value("k", "batchTest2").value("t", "val2");
+        session.execute(batch().add(simple).add(built));
+
+        List<Row> rows = session.execute(select().from(TABLE1).where(in("k", "batchTest1", "batchTest2"))).all();
+        assertEquals(2, rows.size());
+
+        Row r1 = rows.get(0);
+        assertEquals("batchTest1", r1.getString("k"));
+        assertEquals("val1", r1.getString("t"));
+
+        Row r2 = rows.get(1);
+        assertEquals("batchTest2", r2.getString("k"));
+        assertEquals("val2", r2.getString("t"));
     }
 }